# frontend-next/.openapi-cache.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Learning Roadmap Generation System",
    "description": "基于 Multi-Agent 的个性化学习路线图生成系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/roadmaps/{roadmap_id}/retry-failed": {
      "post": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Retry Failed Content",
        "description": "断点续传：重新生成失败的内容\n\n精确到内容类型粒度，只重跑失败的部分：\n- content_status='failed' 的教程\n- resources_status='failed' 的资源推荐\n- quiz_status='failed' 的测验\n\nArgs:\n    roadmap_id: 路线图 ID\n    request: 包含用户偏好和要重试的内容类型\n    \nReturns:\n    task_id 用于 WebSocket 订阅进度",
        "operationId": "retry_failed_content_api_v1_roadmaps__roadmap_id__retry_failed_post",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__api__v1__roadmap__RetryFailedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/roadmaps/{roadmap_id}/concepts/{concept_id}/regenerate": {
      "post": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Regenerate Tutorial",
        "description": "重新生成指定概念的教程（创建新版本）\n\nArgs:\n    roadmap_id: 路线图 ID\n    concept_id: 概念 ID\n    request: 包含用户 ID 和学习偏好\n    \nReturns:\n    新生成的教程元数据（包含版本号）\n    \n说明：\n- 重新生成会创建新版本，不会覆盖旧版本\n- 新版本会自动标记为最新（is_latest=True）\n- 旧版本会被标记为非最新（is_latest=False）",
        "operationId": "regenerate_tutorial_api_v1_roadmaps__roadmap_id__concepts__concept_id__regenerate_post",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          },
          {
            "name": "concept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Concept Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateTutorialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/retry": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Retry Task",
        "description": "智能重试失败的任务\n\n两种重试策略：\n1. **Checkpoint 恢复**（粗粒度）：\n   - 适用于工作流早期阶段失败（intent_analysis, curriculum_design, roadmap_edit等）\n   - 从 LangGraph checkpoint 恢复，继续执行工作流\n   - 保留所有中间状态\n\n2. **内容重试**（细粒度）：\n   - 适用于内容生成阶段部分失败\n   - 扫描 framework_data，只重新生成失败的 Concept 内容\n   - 节省 API 调用成本\n\n自动判断逻辑：\n- 如果存在 checkpoint 且任务失败在内容生成之前 → Checkpoint 恢复\n- 如果有失败的 Concept 内容 → 内容重试\n- 如果 force_checkpoint=true → 强制使用 Checkpoint 恢复\n\nArgs:\n    task_id: 任务 ID\n    user_id: 用户 ID\n    force_checkpoint: 强制使用 checkpoint 恢复\n    background_tasks: FastAPI 后台任务\n    db: 数据库会话\n    \nReturns:\n    重试任务的信息\n    \nRaises:\n    HTTPException: 404 - 任务不存在\n    HTTPException: 400 - 无法重试（没有 checkpoint 也没有失败内容）",
        "operationId": "retry_task_api_v1_tasks__task_id__retry_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "用户ID",
              "title": "User Id"
            },
            "description": "用户ID"
          },
          {
            "name": "force_checkpoint",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "强制使用 checkpoint 恢复",
              "default": false,
              "title": "Force Checkpoint"
            },
            "description": "强制使用 checkpoint 恢复"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LearningPreferences": {
        "properties": {
          "learning_goal": {
            "type": "string",
            "title": "Learning Goal",
            "description": "学习目标，如'成为全栈工程师'"
          },
          "available_hours_per_week": {
            "type": "integer",
            "maximum": 168,
            "minimum": 1,
            "title": "Available Hours Per Week",
            "description": "每周可投入小时数"
          },
          "motivation": {
            "type": "string",
            "title": "Motivation",
            "description": "学习动机，如'转行'、'升职'、'兴趣'"
          },
          "current_level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "title": "Current Level",
            "description": "当前掌握程度"
          },
          "career_background": {
            "type": "string",
            "title": "Career Background",
            "description": "职业背景，如'市场营销 5 年经验'"
          },
          "content_preference": {
            "items": {
              "type": "string",
              "enum": [
                "visual",
                "text",
                "audio",
                "hands_on"
              ]
            },
            "type": "array",
            "title": "Content Preference",
            "description": "偏好的内容类型",
            "default": [
              "visual",
              "text"
            ]
          },
          "target_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Deadline",
            "description": "期望完成时间"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry",
            "description": "所属行业"
          },
          "current_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Role",
            "description": "当前职位"
          },
          "tech_stack": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tech Stack",
            "description": "技术栈列表"
          },
          "preferred_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Language",
            "description": "偏好的学习语言（向后兼容）"
          },
          "primary_language": {
            "type": "string",
            "title": "Primary Language",
            "description": "主要语言（教程、路线图语言）",
            "default": "zh"
          },
          "secondary_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secondary Language",
            "description": "次要语言（资源推荐补充语言）"
          }
        },
        "type": "object",
        "required": [
          "learning_goal",
          "available_hours_per_week",
          "motivation",
          "current_level",
          "career_background"
        ],
        "title": "LearningPreferences",
        "description": "学习偏好配置"
      },
...
      "RegenerateTutorialRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "用户 ID"
          },
          "preferences": {
            "$ref": "#/components/schemas/LearningPreferences",
            "description": "用户学习偏好"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "preferences"
        ],
        "title": "RegenerateTutorialRequest",
        "description": "重新生成教程请求"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__api__v1__roadmap__RetryFailedRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "用户 ID"
          },
          "preferences": {
            "$ref": "#/components/schemas/LearningPreferences",
            "description": "用户学习偏好"
          },
          "content_types": {
            "items": {
              "type": "string",
              "enum": [
                "tutorial",
                "resources",
                "quiz"
              ]
            },
            "type": "array",
            "title": "Content Types",
            "description": "要重试的内容类型列表",
            "default": [
              "tutorial",
              "resources",
              "quiz"
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "preferences"
        ],
        "title": "RetryFailedRequest",
        "description": "断点续传请求"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/jwt/login"
          }
        }
      }
    }
  }
}